Restructured entire CSS template / organized
Political was originally three different attempts at a theme spliced together. As a result it had many duplicated selectors and elements and there was no organization to the overall theme files.

Using Super.so's list of Notion blocks that Super supports it was simple to create the outline and fit like with like to create a final structured file that made sense.

VARIABLES are at the top.

CUSTOMIZATIONS are beneath the variables. CUSTOMIZATIONS ONLY DEAL WITH CHANGES TO FONT, SPACING, PADDING, ETC.
OPTIMIZATIONS include any property which included a `display: block` or `display:none` setting which could be turned on or off.

MOBILE RESPONSIVE is the last section. It reduces the viewport and creates a more readable digestible layout of the theme on mobile devices.