This won't work.
This won't work but need to see how it is handled.

The Callout Button issue lies in how I am calling it. The [class="notion-callout bg-color-light border"] is an outer element that is causing me all the problems.

Maybe.. [class="notion-callout notioncallout__content:first-of-type > notion-emantic-string span notion-link"] will work instead.