broke everything..
reverting back to --color variables with straight RGB (no colorize reference) because it broke everything.

This makes it more difficult to code without constantly seeing the site but to make up for it I set the colors next to the --color variables.

not as good as colorize though. this sucks.